+++ src/backend/Core/Core.Repositories/Interfaces/IUserRoleRepository.cs
@@ -5,7 +5,7 @@ namespace Core.Repositories.Interfaces
     public interface IUserRoleRepository
     {
         Task<bool> IsUserRoleExist(long userId, int roleId);
-        Task AddUserRoleAsync(long userId, long roleId);
+        Task AddUserRoleAsync(long userId, int roleId);
         Task<List<Role>> GetUserRolesAsync(long userId);
     }
 }
